Reorganisation des routes autour du layout
