--- FrontEnd/src/pages/Header.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import 'bootstrap/dist/css/bootstrap.min.css';
-const Header: React.FC = () => {
-    return (
-        <header className="my-4 text-center">
-            <h1>Welcome to SportsSquad</h1>
-        </header>
-    );
-};
-
-export default Header;
