+++ FrontEnd/src/routes/routes.tsx
@@ -1,13 +1,15 @@
 import { Route, Routes } from 'react-router-dom';
-import HomePage from '../pages/Home';
-import ConditionsUtilisation from "../pages/ConditionsUtilisation.tsx";
-import SearchPage from "../pages/Search.tsx";
-import ProfilUser from '../pages/ProfilUser';
+import HomePage from '../Components/HomeContent.tsx';
+import ConditionsUtilisation from "../Components/ConditionsUtilisation.tsx";
+import SearchPage from "../Components/Search.tsx";
+import ProfilUser from '../Components/ProfilUser.tsx';
 
 const routes = () => {
     return (
         <Routes>
             <Route path="/" element={<HomePage />} />
+            <Route path="/search" element={<SearchPage />} />
+            <Route path="/profil/:id" element={<ProfilUser />} />
             <Route path="/conditions-utilisation" element={<ConditionsUtilisation />} />
         </Routes>
     );
